Introduce gbp-buildpackage-rpm tool
Initial version of gbp-buildpackage-rpm - tool for building RPM packages
from git. Bigger differences to the (Debian) gbp-buildpackage are that
buildpackage-rpm always does the build under a separate build directory
(../rpmbuild by default) and that packaging files of a non-native
package may be maintained in an orphan branch, without any sourcecode.
Also, whether the package is native or not is determined by the
existence of an upstream branch or from an explicit config setting.

Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>
Signed-off-by: Ed Bartosh <eduard.bartosh@intel.com>
Signed-off-by: Zhang Qiang <qiang.z.zhang@intel.com>