config: convert to using argparse
Utilize the newer and more flexible argparse module instead of optparse.
Configuration file loading is handled in a separate class.

This patch implements a gbp-specific argparser class for handling prefix
and using the config as default values. The parser is actually a wrapper
class around argument parser and argument groups in order to avoid
inheriting private classes of the argparse module. All commands are
changed to use the newly added gbp-specific argparser.

Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>